requirements.txt
pandas=1.3.5
mysql-connector-python=8.0.27
boto3=1.20.26
python-dotenv=0.19.2
